databases/apq-mysql: Support default MySQL (5.6)
This port generates some headers based on MySQL headers, and a couple
of the variables must have been deprecated because the resulting
variable name ended in "__UNUSED".  Double underscores are not allowed
in Ada, so the build broke.

Since the Ada header is generated during the configure phase, I'm adding
a post-configure target to replace "__" with "_" to restore the build.

Since the dependencies change from MySQL 5.5 to MySQL 5.6, revbump.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@380375 35697150-7ecd-e111-bb59-0022644237b5