Stop forcing the port version string into the server banner.
The port now uses VersionAddendum in the sshd_config to allow overriding
this value. Using "none" allows disabling the default of the port
version string. The default is kept to show the port version string to
remain close to the base version.

Support for the client VersionAddendum may be added soon as well to better
match base and not give surprises when switching from base to the port.

PR:		193127
Requested by:	many, including myself when this was broken years ago.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@381981 35697150-7ecd-e111-bb59-0022644237b5