New port: net/rdapper
rdapper is a command-line client for the Registration Data Access
Protocol (RDAP), as described in the RDAP RFCs (7480 - 7485).

This tool will send an RDAP query to an RDAP server over HTTP or
HTTPS, parse the JSON response, and display it in human-readable
form.

WWW: https://github.com/jodrell/rdapper

Submitted by:	pi


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@382343 35697150-7ecd-e111-bb59-0022644237b5