devel/libc++: Mainly fix DragonFly support.
This port was originally fixed in dports using alias:10 but this fix
was faulty.  The libc++ headers reference __FreeBSD__ so alias won't
work.  It needs to be patched to check __DragonFly__, including a file
that was the subject of EXTRA_PATCHES.

Moreover, the port was installing the .orig files left over from
patching.  I remove these in post-patch now.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@383184 35697150-7ecd-e111-bb59-0022644237b5