Chase the xen-tools -> xen-guest-tools rename
xe-guest-utilities needs to use this older version of the xen tools
which are compatible with FreeBSD 9.x and 10.x. The tools are required
as part of the xe-guest-utilities scripts to report data back to the
hypervisor and permit additional functionality such as Live Migration
when running FreeBSD VMs on the Citrix XenServer hypervisor.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@383456 35697150-7ecd-e111-bb59-0022644237b5