cad/ldraw: Fix fetch; distfile was rerolled
I have the idea that ldraw is always released under the same filename.
The port was marked broken due to a checksum mismatch; the file was
fetched but it had changed.

I've expanded DIST_SUBDIR to include the PORTVERSION.  I think the scheme
used is just {YEAR}{ITERATION-2-DIGIT} so I picked 201501 as the new
PORTREVISION.  This builds under poudriere testport just fine.

Approved by:	blanket


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@383822 35697150-7ecd-e111-bb59-0022644237b5