- Add p5-VM-EC2 1.28
This is an interface to Amazon AWS API. It was written provide access to the new
tag and metadata interface that is not currently supported by Net::Amazon::EC2,
as well as to provide developers with an extension mechanism for the API. This
library will also support the Open Stack open source cloud.

The main interface is the VM::EC2 object, which provides methods for
interrogating the Amazon EC2, launching instances, and managing instance
lifecycle. These methods return major object classes which act as specialized
interfaces to AWS.

WWW: http://search.cpan.org/dist/VM-EC2/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@384015 35697150-7ecd-e111-bb59-0022644237b5