Remove USES=compiler:c++11-lib from mail/evolution.  While it fixes the build failure on FreeBSD 8 (and probably 9), the application fails to start because the base version of libstdc++ is loaded at runtime, which does not satify the requirements of webkit-gtk3.  The reason is that evolution does not directly link to libstdc++, so the rpath added to it by USES=compiler:c++11-lib has no effect.  The first shared library that links to libstdc++ belongs to evolution-data-server, which does not have USES=compiler:c++11-lib, which causes the base version of libstdc++ to be loaded.
Unconditionally add USES=compiler:c++11-lib to
databases/evolution-data-server.  Adding the rpath to its shared
libraries will cause the newer version of libstdc++ to be loaded
when evolution starts and fix the runtime error on FreeBSD 8 and
9.  For standard versions of FreeBSD 10 and higher on amd64 and
i386, it is a no-op, so there is no harm in doing this. If FreeBSD
10 or higher is configured to build with base gcc and clang is not
present, which is an option on amd64 and i386 and is mandatory on
some tier 2 platforms that don't have clang support, then
USES=compiler:c++11-lib is needed.

PR:		199746
Differential Revision:	https://reviews.freebsd.org/D2388
Reviewed by:	kwm
Approved by:	mat (mentor)
MFH:		2015Q2


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@384997 35697150-7ecd-e111-bb59-0022644237b5