textproc/dblatex: Fix failure to invoke, reset maintainer
The shebang "env python" won't work.  Normally shebangfix would be used
to fix this, but dblatex is actually built so we need to use sed in the
post-install target to adjust it.

While here, reset the maintainer.  This PR is the fourth consecutive
timeout.  Also clean up the port.

PR:		196617
Submitted by:	holger (freyther.de)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@386630 35697150-7ecd-e111-bb59-0022644237b5