benchmarks/pathrate: Unbreak by setting new MASTER_SITES
While here, set the license (GPLv2).  The other changes to the Makefile
and the non-WWW part of pkg-descr are technically unnecessary, but in
my opinion they are in improvement so I keep them.  I did remove some
of the suggested changes in the patch though.

PR:		200295
Submitted by:	Ports Fury


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@386751 35697150-7ecd-e111-bb59-0022644237b5