emulators/qmc2: Adjust after update to mame/mess
The mame and mess executable names have changed slightly.  For i386
platforms, they are still "mame" and "mess" but for amd64 platforms,
they are "mame64" and "mess64".  Change the RUN_DEPENDS specification
to adjust for this morphing name.

I suspect a patch is also needed for amd64 but I don't know where it
would go.  I'll notify the maintainer.  The RUN_DEPENDS fix and the
hardcoded CFLAG fix are "just fix it" items.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@387375 35697150-7ecd-e111-bb59-0022644237b5