x11/hotwire-shell: drop pyxml dependency
This port depends on prehistoric pyxml library, but it doesn't using it
actually. Here is two modules that it's using:

>>> import xml.sax, xml.sax.handler
>>> from xml.sax.saxutils import escape

Both of them are in python standard library for quite some time, so this
dependency can be just dropped altogether.

While here, remove redundancy in COMMENT, limit python version to 2.x
because of pygtk2, sort USE_PYTHON components and remove not-functioning
WWW link.

Approved by:	ashish (maintainer, by email)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@388983 35697150-7ecd-e111-bb59-0022644237b5