Update PostgreSQL to latest versions.
Earlier update releases attempted to fix an issue in PostgreSQL 9.3 and 9.4
with "multixact wraparound", but failed to account for issues doing multixact
cleanup during crash recovery. This could cause servers to be unable to restart
after a crash. As such, all users of 9.3 and 9.4 should apply this update as
soon as possible.

URL:	http://www.postgresql.org/about/news/1592/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@389267 35697150-7ecd-e111-bb59-0022644237b5