- Fix build on CURRENT
In base revision r281198, sockaddr_l2cap structure was changed, and
an #error which fires by default was added as a safety measure to
make sure all structure consumers are updated to initialize new
fields. This port, however, doesn't seem to use this structure, so
it should be safe to just disable the error.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@389989 35697150-7ecd-e111-bb59-0022644237b5