Add devel/cloudabi-cloudlibc.
CloudABI uses a C library called cloudlibc. It's similar to a standard
UNIX C library, except that all features that use global namespaces have
been removed. The advantage is that this causes compiler errors when
trying to use existing software, instead of causing it to fail in
unpredictable ways at runtime.

This package installs a copy of cloudlibc for CloudABI/x86-64 (the only
supported platform for now). It is installed in the right place, so that
Clang and Binutils can find the headers/library archives.

Differential Revision:	https://reviews.freebsd.org/D2886
Approved by:	bapt


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@390359 35697150-7ecd-e111-bb59-0022644237b5