Add devel/cloudabi-compiler-rt.
This is the last port that is needed to have a working C toolchain for
CloudABI. C++ still requires libc++, libc++abi and libunwind. These will
be added later.

As mentioned in the comments, we don't use CMake to build compiler-rt,
as this is too hard in this specific instance. CMake sort of depends on
compiler-rt to be present, which is a bit weird.

Differential Revision:	https://reviews.freebsd.org/D2892
Reviewed by:	bapt


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@390444 35697150-7ecd-e111-bb59-0022644237b5