www/py-django-dpaste: Update to 2.8
- Update to 2.8
- Switch to CHEESESHOP for MASTER_SITES
- Match COMMENT with upstream
- Add LICENSE_FILE
- Sort USE_PYTHON values
- Tweak RUN_DEPENDS to limit to Django 1.6 and switch to file-based
  django detection (instead of package).

This ensures that if a port that depends on this one, also depends on a
different django version, it will see and use the installed version,
rather than attempting to install another one, which results in a
conflict as all django ports conflict with each other.

See r390184 [1] and 390535 [2] for other instances of this issue.

- Remove TESTS option and tox as a TEST_DEPENDS

[1] https://svnweb.freebsd.org/changeset/ports/390184
[2] https://svnweb.freebsd.org/changeset/ports/390535


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@390536 35697150-7ecd-e111-bb59-0022644237b5