Require python as a run dependency, too
python is required for you to be able to build custom vmods and also for
other vmods in the ports tree to build. Better off including the
requirement here as we ship the vmodtool.py utility and do not want to
surprise end users.

shebangfix was included as the default shebang on vmodtool.py is
/usr/bin/env python which is not going to work unless the lang/python
metaport is installed, which is not included with USES=python. Instead
we just rewrite this to whichever the default system python intepreter
is (PYTHON_VERISON).


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@390797 35697150-7ecd-e111-bb59-0022644237b5