databases/p5-Memcached-libmemcached: improve patch
The patch to Makefile.PL needed to also replace $Config{lib_ext} (which
defaults to ".a") with ".so" so that we link against the right thing. This
avoids the need to patch databases/libmemcached.

While here, also avoid hard coding the path to PREFIX and split out the
libstdc++ to libc++ patch, avoiding duplication.

Pointyhat to:		swills
Reported by:		tijl, bapt


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@391019 35697150-7ecd-e111-bb59-0022644237b5