Update tryton 2.4 suite to it's ultimate versions
Tryton 2.4 is no longer supported by the upstream project. Users are encouraged
to upgrade to a newer version as soon as possible (we have tryton 2.8 in ports).
So update this ports to it's ultimate versions and mark it deprecated.
Set expiration date to 2015-08-05.

PR:		197935
Submitted by:	Matthias Petermann <matthias@petermann-it.de> (maintainer)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@391349 35697150-7ecd-e111-bb59-0022644237b5