cad/tkgate: Unbreak FreeBSD 11, Fix depends on rest
This was specified on USES: iconv:build,lib

This is incorrect; the way iconv.mk is written, "build" and "lib" are
mutually exclusive and "build" takes precedence.  This means the library
dependency was not registered.

Moveover, it's failing on FreeBSD 11 and dev-branch of DragonFly because
it uses transliteration.  Setting USES=iconv:translit fixes both issues.

Approved by:	Just fix it


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@391493 35697150-7ecd-e111-bb59-0022644237b5