graphics/skencil: mark deprecated
This is quite old project being abandoned: last release was in year 2010. It's
code base relying on stale libraries and in fact it was removed from most of
the distributions. Users should switch to graphics/inkscape or another mature
vector editor. There is also Skencil's fork - sK1 [1] that is actively
developing, but that we have not in ports tree yet. There is also some story
behind the fork [2].
Not sure if this a problem of this app in gnome3 environment, or some general
failure, I wasn't able to add simple text label to the canvas getting this
everywhere:

"""
EditSelection.GetObjectMethod: editor <Sketch.Graphics.selection.EditorWrapper
instance at 0x80a128560> is not compatible
with class skencil.Sketch.Graphics.text.CommonTextEditor
"""

Mark this deprecated and set expiration date to 2015-08-08.

[1] http://sk1project.org/news.php?readmore=107
[2] http://sk1project.org/viewpage.php?page_id=21


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@391598 35697150-7ecd-e111-bb59-0022644237b5