Add sysutils/cloudabi-utils.
Summary:
Though CloudABI programs can simply be started from the command line
like any UNIX process, the preferred way for starting them is using the
cloudabi-run utility.

cloudabi-run parses a YAML file that contains the configuration for the
program that you want to run. Special tags can be used in the YAML file
to indicate that the program depends on a resources (a socket or a
file). cloudabi-run then attempts to obtain these resources and startes
the program for you.

The end result is that it's possible to build software that looks like
traditional UNIX software (executable + config file), but still gets run
in this completely sandboxed environment.

Reviewers: bapt

Reviewed By: bapt

Differential Revision: https://reviews.freebsd.org/D3067


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@391894 35697150-7ecd-e111-bb59-0022644237b5