games/galaxymage: mark broken and deprecated
Game fails to start with this message:
IOError: [Errno 2] No translation file found for domain: 'GalaxyMage'

Locale stuff in this app is really weird, and I wasn't able to fix it
in affordable time, so just mark it broken, because in it's current
state it completely unusable. This project was abandoned upstream
since 2006 and then later revilitalized under slightly changed name,
but no releases were issued since 2011, when this fork started.

If anybody willing to fix this port, lets get in contact - I also have
patches to make it work with numpy instead of Numeric.

Mark deprecated, set expiration date to 2015-08-16.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@392321 35697150-7ecd-e111-bb59-0022644237b5