New port: sysutils/reptyr
reptyr is a utility for taking an existing running program and
attaching it to a new terminal. Started a long-running process over
ssh, but have to leave and don't want to interrupt it? Just start
a screen, use reptyr to grab it, and then kill the ssh session and
head on home.

WWW: http://github.com/nelhage/reptyr

PR:		200720
Submitted by:	 Andrey Cherkashin <andoriyu@gmail.com>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@392394 35697150-7ecd-e111-bb59-0022644237b5