[Forced commit to explain the previous commit]
Before this, you could say:
BUILD_DEPENDS=	libfoo.so:some/foo
LIB_DEPENDS=	bar:some/bar

The package would build because the dependencies would be installed
correctly.  But, the dependencies would not be registered in the
package, and it would be non functionnal when installed.

Now, if you say the same thing, it will tell you that
1) it can't find exececutable libfoo.so
2) bar is not a valid pattern for LIB_DEPENDS

giving you a hint that you are doing something wrong.

Poked by:	myself, bapt


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@392818 35697150-7ecd-e111-bb59-0022644237b5