Remove cross compiled libraries for CloudABI.
These libraries are now provided by the CloudABI Ports collection -- not
just for FreeBSD, but a number of other operating systems. For FreeBSD,
they are provided in the form of a pkg mirror, so there is no need to
also have them in the FreeBSD Ports tree.

Approved by:	bapt
Differential Revision:	https://reviews.freebsd.org/D3643


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@396738 35697150-7ecd-e111-bb59-0022644237b5