lang/mono: Remove $FreeBSD$ from patches
Patches aren't supposed to have $FreeBSD$ tags anymore.  The
fbsd:nokeywords property was properly set, so the $FreeBSD$ tag was not
even expanded, so these tags were removed.  Note that the mini-exceptions
patch is no longer needed.  According to the comments, it's is for
FreeBSD 8 support only.

Approved by:	portmgr (mat)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@397063 35697150-7ecd-e111-bb59-0022644237b5