MFH: r396461
It turned out -fPIC is required on sparc64 as well: it dumps core without
it on extraction, which turn breaks e.g. `sysutils/namefix' port.  Do not
bump port revision again as it was bumped just couple of hours ago, which
is just too small in Tier-2 land for which we do not produce any packages
anyways.

Tested on:	flame
Approved by:	ports-secteam (with hat)