devel/ahven: Upgrade version 2.4 => 2.6
I could not use the comfignat build system.  For starters it creates
a dynamic library which, when used, has troubling finding libgnarl.
If I make changes to force it to build statically, the programms that
link to it fail with a STORAGE ERROR.

Finally I just put all the build and install steps in the port makefile
and everything is working as it did with version 2.4


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@401003 35697150-7ecd-e111-bb59-0022644237b5