Simplify Makefile and use the right flags when building.
Remove the custom do-configure and do-build target. The targets existed
because upstream's build instructions mention external.pro and
meshlab_full.pro need to be processed and built separately and Uses/qmake.mk
does not handle the multi-.pro case. do-build, however, was simply calling
${MAKE} and discarding both ${MAKE_ENV} and ${MAKE_ARGS} which, among other
things, resulted in no parallel builds and additional compiler flags like
-std=c++11 being omitted.

Instead, patch meshlab_full.pro and add external/external.pro as one of the
directory dependencies there along with the rest of the MeshLab code. Given
the way the QMake files are written in the port, this only helps a bit with
parallelism: each subdirectory is processed serially, but all files in them
can be built in parallel. Since a lot of the code is in plugin directories
with 1 or 2 files, the port still takes some time to build. The upside is
that it means external/ will always be built before the rest, which is what
we need.

As mentioned, the compiler flags in the build will change. However, I'm not
bumping PORTREVISION yet because it will come with the next commit updating
math/mpir.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@402166 35697150-7ecd-e111-bb59-0022644237b5