x11-fm/doublecmd: Make less broken, then set BROKEN
After the upgrade to FPC 3.0.0, the port failed the build depends stage.
I've fixed that, and it compiles pretty far until this error:

  doublecmd.lpr(18,3) Fatal: (10022) Can't find unit uOverlayScrollBarFix

I suspect that is a Lazarus unit and it might be a missing path.  In the
meantime, the port can't build so we have to mark it BROKEN.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@404187 35697150-7ecd-e111-bb59-0022644237b5