devel/ada-util: Add many missing source files, remove xmlada support
I have my doubts about gprinstall.  It missed installing many source
files.  I can only assume it's buggy.  Scoop them up in a crude fashion
with post-install target.

While here, turn off xmlada support.  The configure script in this are
is hopelessly broken (it doesn't want to accept no for an answer) so
just rip out the configure script guts.  XML/Ada is GPL and it will
carry the license to any project that uses Ada-util.  It also makes
XML/Ada a run dependency for the same project.  Probably in the future
I will return support in form of a non-default option, but I'll want
up to a more recent version.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@404784 35697150-7ecd-e111-bb59-0022644237b5