multimedia/plexmediaserver(-plexpass): Fix scanner crash
The Plex processes were running under the C locale. This was to
replicate the behavior of the start.sh shell script shipped with Plex and
used by Linux distributions. The scanner process was crashing when it
encountered unicode filenames due to this locale.
Upstream Plex fixed this, but we were left affected.

This resolves the scanner process crash on FreeBSD.

PR:		205696


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@404948 35697150-7ecd-e111-bb59-0022644237b5