MFH: r405358
Fix MASTER_SITES, regenerate distinfo and unmark BROKEN.

The old download URL was pointing to the zip version of the source code. Switch
to the url that serves a .tar.gz, regenerate distinfo (it contained information
about the .zip file with the wrong extension) and remove a non-working mirror.

Approved by:   portmgr blanket