vietnamese/vietunicode-*: Fix options (3 ports)
The -hannom, -web1, and -trichlor ports all had their options mis-set
by the unnecessary use of <pre> and <post>.  They sandwhiched a third
makefile that pulled in <options> which is sufficient.  Remove <pre>
and <post> inclusions to restore option functionality of these ports.
Bump to be on the safe side.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@406308 35697150-7ecd-e111-bb59-0022644237b5