ports-mgmt/synth: v0.99 (RC), fixes / use official pkgs feature
Now Synth has all the features envisioned for the first release.  This
edition add an option (off by default) to fetch prebuilt packages if
they are suitable (ABI, options, dependencies match).  This feature is
aimed at people that only want to build ports with customized options,
but for ports with default configurations, they are happy to use the
official packages.  The feature is not heavily tested yet.

Once this version is sufficently tested, Release 1.00 will follow.

Other changes since 0.98_5:
  * curses will redrawn itself (correctly now) every 30 seconds to
    fix any corruption that may have occured
  * Synth everything will build the repository without deleting packages
    now (this was disabled on the last update)
  * The repos directory for pkg is read from pkg config instead of using
    the hardcoded defaults
  * The man page has been updated with new feature descriptions
  * linprocfs is mounted for linux_base ports
  * The regression that prevented linux ports from building has been
    fixed.  It was caused by caching LINUX_OSVERSION, so this variable is
    no longer cached.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@407249 35697150-7ecd-e111-bb59-0022644237b5