[NEW] x11-themes/gtk-E17-theme: Dark GTK theme for Enlightenment
E17gtk is a dark Gtk theme with sharp corners, which is designed for use in
Enlightenment and gives the elegant look of Enlightenment to Gtk3, Gtk2 and
Qt widgets. Of course, it can be used with any GTK based environment too,
but it is not tuned to Gnome or any of its apps.

WWW: http://gnome-look.org/content/show.php/?content=163472

PR:		205475
Submitted by:	Matthew <matthew reztek cz>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@407831 35697150-7ecd-e111-bb59-0022644237b5