ports-mgmt/synth: Fix end-run regression
A fairly recent change caused a regression after a build was complete.
Previously a "tally" or summary of the build would appear after the
ncurses screen was restored to the regular terminal mode.  It would
list how many ports were built, failed, etc.  After the regressin, it
just ended abruptly.

This commit restores the tally to show as it did previously.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@408035 35697150-7ecd-e111-bb59-0022644237b5