Lighten up the Perl dependencies.
git uses Perl for two things: Perl hooks into git, and the
git-send-email(1) script.

The Perl hooks only use p5-Error. The other modules dependencies,
p5-Authen-SASL and p5-Net-SMTP-SSL (which bring in a number of other
perl module dependencies) are only required for git-send-email(1).

This commit adds a SEND_EMAIL option, defaulted to on, that auto-enables
the PERL option and installs the git-send-email(1) script with the
extra perl modules.

With the PERL option on and SEND_EMAIL off, only the p5-Error module
is required.

No PORTREVISION bump as the default dependencies and plist haven't changed.

PR:		206901
Approved by:	garga (maintainer)
Differential Revision:	https://reviews.freebsd.org/D5179


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@408063 35697150-7ecd-e111-bb59-0022644237b5