science/py-pysal: Avoid interactive build possibility
This port checks the release date, and subtracts it from the current
date.  If the difference is more than 6 months, it interactively asks
the user if they want to upgrade.

The post-patch target removes this possibility, ensuring the port never
reverts to an interactive build.

approved by:	infrastructure blanket (interactive builds discouraged)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@408407 35697150-7ecd-e111-bb59-0022644237b5