New port: games/linux-unigine-heaven
Heaven Benchmark immerses a user into a magical steampunk world of
shiny brass, wood and gears. Nested on flying islands, a tiny village
with its cozy, sun-heated cobblestone streets, an elaborately crafted
dirigible above the expanse of fluffy clouds, and a majestic dragon
on the central square gives a true sense of adventure. An interactive
experience with fly-by and walk-through modes allows for exploring
all corners of this world powered by the cutting-edge UNIGINE Engine
that leverages the most advanced capabilities of graphics APIs and
turns this benchmark into a visual masterpiece.

This port includes the linux-binaries of the basic (free-to-use)
version of the benchmark.

WWW: http://unigine.com/products/heaven/

PR:		194472
Submitted by:	Hannes Hauswedell <h2+fbsdports@fsfe.org>
Reviewed by:	marino, d.akiyakov@gmail.com


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@408535 35697150-7ecd-e111-bb59-0022644237b5