New port: devel/py-arrow
Arrow is a Python library that offers a sensible, human-friendly approach
to creating, manipulating, formatting and converting dates, times,
and timestamps. It implements and updates the datetime type, plugging gaps
in functionality, and provides an intelligent module API that supports
many common creation scenarios. Simply put, it helps you work with dates
and times with fewer imports and a lot less code.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@409363 35697150-7ecd-e111-bb59-0022644237b5