Remove russian/muttprint.
The port was created in 2003 and never updated again except when it needed
to be adapted to changes in the ports framework itself. It has not had a
maintainer since 2007.

After updating print/muttprint to 0.73, the translation was not even
accurate anymore, and upstream already includes an updated version of the
same file (translation-ru.pl). The port was also broken after the
print/muttprint update, as we stopped setting variables such as CATEGORIES,
COMMENT and MAINTAINER with "?=" instead of "=".

PR:		207299


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@409522 35697150-7ecd-e111-bb59-0022644237b5