lang/itcl4: update to 4.0.4
Itcl 4.0.4 fixes a large collection of bugs and flawed designs. See the ticket
systems and timelines for a full accounting.  A great deal of progress has been
made to make Itcl 4 a functionally equal replacement for Itcl 3.  If you've
delayed migration due to lack of compatibility, it's time to try again.  This
is the same Itcl release that is bundled with Tcl 8.6.5.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@409929 35697150-7ecd-e111-bb59-0022644237b5