MFH: r410039
security/openssl: Revert disabling of SSLv2 and MD2

Disabling SSLv2 without a shared library bump has a visible impact to
some applications. It is unclear at this time if disabling MD2 could
cause the same issues, but both are being reverted at the moment to be
safe.

PR:		195796
Approved by:	ports-secteam (with hat)