net/guacamole-server, misc/e2fsprogs-libuuid, misc/p5-OSSP-uuid, misc/ossp-uuid
Currently when requiring both ossp-uuid and e2fspogs-libuuid is an
impossibility without the use of a jail. This patch was created to
fix this conflict. The patch includes updating revisions for removal
of CONFLICTS, and change LIBDEPEND change for all affected ports.

The changes on ossp-uuid port includes renaming of the shared/static
libs created, C/C++ manpage, and rename of pkgconfig file.

PR:		207004
Submitted by:	Ultima1252@gmail.com
Approved by:	vd


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@410454 35697150-7ecd-e111-bb59-0022644237b5