ports-mgmt/synth: Upgrade version 1.20 => 1.21
I knew the repository signing enhancements weren't fully tested, but it
was important to get the regressions fixed.  I thought that if there were
any problems with the signing, people not using those features (e.g. 99%
of users) would not be affected.  Unfortunately, that's not true.

Due to a simple boolean logic error, an inaccurate and harmless warning
about not finding the signature fingerprint will show every time the
repository is rebuild.  This minor releases fixes that logic.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@410606 35697150-7ecd-e111-bb59-0022644237b5