qt4 sqldrivers: Switch to a proper qmake build, use Makefile.sqldrivers.
We've recently started receiving pkg-fallout emails because qt4-mysql-plugin
is failing to build in HEAD. It turns out we were using some custom-made
Makefile.bsd files to drive the builds, and they did not always register all
dependencies between the files correctly.

Fix it by switching to a proper qmake build that uses the .pro files shipped
with Qt4 itself: they can be used without running the `configure' script
almost as if they were not part of the Qt distribution itself. By doing this
we can stop having our own Makefiles and also stop setting a lot of
variables in the port Makefiles.

While here, consolidate most of the variable setting into a single
Makefile.sqldrivers in devel/qt4 (like we already do for devel/qt5) so that
each of the qt4-*-plugin ports only need to set a few values such as the
plugin name and additional USES or includes that might be necessary.

Bump PORTREVISION because we now include the debug versions of the plugins
in PLIST_FILES when the ports are built with WITH_DEBUG=yes (they were
already shipped before, but not registered in the plists).


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@410751 35697150-7ecd-e111-bb59-0022644237b5