New port: archivers/py-rjsmin
rJSmin is a javascript minifier written in python. The minifier is based
on the semantics of jsmin.c by Douglas Crockford. The module is a
re-implementation aiming for speed, so it can be used at runtime
(rather than during a preprocessing step). Usually it produces the same
results as the original jsmin.c.

WWW: https://github.com/ndparker/rjsmin

PR:		208834
Submitted by:	Ultima1252@gmail.com


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@413467 35697150-7ecd-e111-bb59-0022644237b5