[bwn] Initial support for the v5 firmware
This is just to support the initial v5 firmware for one of the PHY-N
NICs I'm going to do some development.

It only pulls out the specific v5 mimo firmware for the BCM4312 NIC
I am playing with; I'm adding it so we at least have the framework
to pull things out of the v5 firmware.

The existing NICs still use v4 firmware and I don't plan on changing
that for some time as bwn(4) doesn't yet support the newer firmware API
and descriptor layouts.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@414669 35697150-7ecd-e111-bb59-0022644237b5