Record TIMESTAMP in make makesum
This is intended to support development and prototyping for ports
reproducible build efforts which require some concept of a "last
updated" time. It is being committed now so that timestamp entries
will be populated "for free" as port distfiles updates happen.

Submitted by:	bapt (earlier version)
Reviewed by:	bapt, bdrewery
Approved by:	portmgr (bdrewery)
Differential Revision:	https://reviews.freebsd.org/D6031


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@415078 35697150-7ecd-e111-bb59-0022644237b5