Fix PolarSSL-based builds.
The upstream backported a change from the master branch that fixes the
PolarSSL-based builds to go with the PolarSSL 1.3.X built-in defaults.

Add a patch picked from the upstream's release/2.3 branch.
Remove the BROKEN= line and conditional.

No PORTREVISION bump because the patch only affects an option that was
formerly marked BROKEN.

(TRYBROKEN users need to force a rebuild and reinstallation manually.)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@415116 35697150-7ecd-e111-bb59-0022644237b5