textproc/py-parsimonious: NEW PORT - PEGs based parser in Python
Parsimonious aims to be the fastest arbitrary-lookahead parser written
in pure Python and the most usable. It's based on parsing expression
grammars (PEGs), which means you feed it a simplified sort of EBNF
notation. Parsimonious was designed to undergird a MediaWiki parser
that wouldn't take 5 seconds or a GB of RAM to do one page, but it's
applicable to all sorts of languages.

WWW: https://pypi.python.org/pypi/parsimonious

PR:		209732
Submitted by:	Danilo G. Baio <dbaio@bsd.com.br>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@415899 35697150-7ecd-e111-bb59-0022644237b5