misc/kde4-xdg-env: fix after kdeprefix removal
Fix SUB_LIST and script

This is related to bug #209799, same root cause: the removal of
kdehier and USE_KDE4=kdeprefix is a bit too thorough. Here, it
leaves us with a blank variable in SUB_LIST, which ends up in a
shell script. The substitution leaves the shell script with syntax
errors.

Since $KDE4_PREFIX must be equal to $LOCALBASE, remove the substitution
and simplify the script.

PR:		209802
Submitted by:	groot@kde.org (kde)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@415977 35697150-7ecd-e111-bb59-0022644237b5