New port: misc/qt5-doc
This port builds and installs the Qt5 API documentation for use
with IDEs such as qtcreator or kdevelop in qch file format and HTML
file format.

Limitations: The port is made for Qt 5.5.1 and excludes the
documentation for qtwebkit, qtwebkit-examples and qtwebengine. The
reasons are, webkit is deprecated and should not be used for further
development, it will not be part of the Qt 5.6.x sources by default
and only provided on FreeBSD for backwards compatibility with ports.
API documentation may be later added to the qt5-webkit port (as
well for qtwebkit-examples). The qtwebengine hasn't been ported to
FreeBSD yet, so no documenation can be generated either.

The build process uses the full qt source tarball and uses configure
with the respective platform (freebsd-g++ or freebsd-clang). During
configure, the qmake tool is compiled that generates the Makefiles
for doc creation. While the qt5-qmake could also be used, it currently
generates the documentation without the correct links and replaces
them with text, which makes the API documentation useless with
respect to the workflow. The compiling of qmake has been accepted
for now as a minor drawback of the port.

PR:		209861
Submitted by:	Ralf Nolden <nolden@kde.org> (kde)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@416213 35697150-7ecd-e111-bb59-0022644237b5