- Reset maintainer by his own request:
"Due to lack of FreeBSD computers, processing facilties and remote
location between now and the end of the year I have to drop
maintainership of this port. If anyone is interested in taking it
on then the routine security updates are usually trivial. I would
hope to be able to take up this port towards the end of the year
but that's not a definite.

More complex would be porting Drupal8 into the ports tree . . . .

Thanks and regards

Simon Wright"

PR:		209998
Submitted by:	simon.wright@gmx.net (maintainer)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@416328 35697150-7ecd-e111-bb59-0022644237b5