databases/cego: 2.30.5 -> 2.30.6
The following query can cause a core dump
  create table t1(a string(100));
  insert into t1 values ( 'Hugo Habicht');
  insert into t1 values ( 'Kim Kaiser');
  select substr(a, 1, getpos(a, ' ')) from t1;

For return value length evaluation, the method CegoFieldValue::asInteger
is called on a null value which causes the dump. A fix has been
added, so the method returns a value of 0 in case of null values.

Submitted by:	Bjoern Lemke <lemke@lemke-it.com>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@416345 35697150-7ecd-e111-bb59-0022644237b5