- Add p5-Test-Needs 0.002002
Skip test scripts if modules are not available. The requested modules will be
loaded, and optionally have their versions checked. If the module is missing,
the test script will be skipped. Modules that are found but fail to compile will
exit with an error rather than skip.

If used in a subtest, the remainder of the subtest will be skipped.

Skipping will work even if some tests have already been run, or if a plan has
been declared.

Versions are checked via a $module->VERSION($wanted_version) call. Versions must
be provided in a format that will be accepted. No extra processing is done on
them.

If perl is used as a module, the version is checked against the running perl
version ($]). The version can be specified as a number, dotted-decimal string,
v-string, or version object.

If the RELEASE_TESTING environment variable is set, the tests will fail rather
than skip. Subtests will be aborted, but the test script will continue running
after that point.

WWW: http://search.cpan.org/dist/Test-Needs/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@417533 35697150-7ecd-e111-bb59-0022644237b5