Upgrade axa to version 1.3.0.
Axa now contains some third party public domain code, add license info for
that.

Respect the chosen default SSL version.  The base version of OpenSSL in
FreeBSD 9 isn't good enough, so mark this port as IGNORE for that case.
Make the port work with the OpenSSL in base for the newer FreeBSD
branches.

Patch the source so that it doesn't install an unnecessary header that
would conflict with converters/libb64, and fix a format string type
mismatch bug.

Sponsored by:	Farsight Security, Inc.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@417911 35697150-7ecd-e111-bb59-0022644237b5