Remove the deinstall script that appears to be deleting all the temporary files created by pkg(8) during upgrades
It happens because the deinstall script tries to clean up the potential manual
VM registration by cleaning out all symlinks to bin/javavm

Given all VM are registring/unregistering themselves this part is not needed

The other thing the script was doing handling the configuration which has been
replaced by @sample.

pkg-install has been modified to drop the handling of the configuration file but
keep the auto registration if all VM found. While this part is not necessary as
well, we keep it because otherwise anyone doing delete/install on javavmwapper
version 2.5 being the installed version would end up with all VM unregistered.

The pkg-install should be removed after EOL of FreeBSD 10.3

PR:		210313
MFH:		2016Q3


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@419364 35697150-7ecd-e111-bb59-0022644237b5