net/fpc-pcap: Remove redundant pcap option
The default LIBPCAP option didn't do anything.  It set a LIB_DEPENDS
that is always satisfied by the base libpcap so no dependencies were
registered.

While here, set PORTREVISION.  The slave ports of fpc should have have
PORTREVISION?= line, even if the value is zero.  This one didn't, so
we have to bump to 2 because the master port already had set it to 1.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@419608 35697150-7ecd-e111-bb59-0022644237b5