MFH: r420300
net-mgmt/rancid3: Fix SSH on FreeBSD 11.0

RANCID by default attempts to ssh to devices with 3des as the cipher.
This no longer works with OpenSSH in FreeBSD 11 as you need to be more
specific. Changing this to 3des-cbc works for now, but future versions
of OpenSSH will break this as well. Unfortunately this will break RANCID
as many network devices do not support modern ciphers, but we will have
to deal with that when the time comes.

Approved by:	ports-secteam (with hat)