Stop trying to build a package
The build process starts by trying to to download nuget and some dependencies
required for building.  It expects them to be downloaded this way, and because
the network is down during build, it fails.  So stop trying to build pakages
for f-sharp: hoping that users interested in f-sharp have networking while
building.
Reported by:	pkg-fallout, poudriere


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@420321 35697150-7ecd-e111-bb59-0022644237b5