multimedia/plexmediaserver: Change the way binaries are stripped
This is a master port with the PlexPass port as the slave. Calling strip
directly on files is problematic when they change or are moved in
updates to the slave port. This is a no-op for this port but prevents
errors in the slave port from trying to call strip on files that no
longer exist or have moved.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@420412 35697150-7ecd-e111-bb59-0022644237b5