graphics/nomacs: Fix build when RAW option turned off
The opencv-core dependency is not limited to the RAW option.  The build
fails in a clean jail if the RAW option is turned off due to the missing
opencv dependency.  Change it to an unconditional LIB_DEPENDS to fix
the build in the RAW-off configuration.

PR:		202497
Reported by:	serpent7776 (gmail)
Approved by:	maintainer timeout / just-fix-it


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@420443 35697150-7ecd-e111-bb59-0022644237b5