devel/poco-devel: Older than stable port, deprecate for removal
The devel version of poco is signficantly behind the stable devel/poco
which is regularly updated.  There's no apparent benefit to having two
ports in this case.  I'll mention to the maintainer that devel/poco is
currently unmaintained in case he wishes to take that over even though
ports fury has been handling that for a while.

PR:	204824


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@420445 35697150-7ecd-e111-bb59-0022644237b5