www/p5-Gtk2-WebKit: Comment out NO_PACKAGE hack from March 2010
It's been 6.5 years since this port was marked with NO_PACKAGE which
should have been a temporary measure.  It builds fine for me on F10.3
amd64 (vm), so let's remove the designation so we can see if it's
still an issue.  If the problem still exists, then it's better make
the documentation optional and off by default.

Approved by:	Just-Fix-It


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@421196 35697150-7ecd-e111-bb59-0022644237b5