net-mgmt/rancid3: Update to 3.5.1
- The hardecoded 3des cipher for ssh has been removed upstream
- Remove a patch that automatically attempts to "svn up" before
checking in changes. This patch may have been committed accidentally a
couple years ago and deviates from upstream's behavior.

If it is possible for your rancid svn repository to be updated outside
of rancid's control it is up to you to resolve these potential
conflicts and modify your scheduled rancid process to "svn up"
beforehand. By instructing rancid to do this automatically it is
possible for an administrator to make device config changes that rancid
will not detect and will not show up in the automatic emails dispatched
to the device administrators.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@421711 35697150-7ecd-e111-bb59-0022644237b5