Ignore SIGINFO during self-tests.
In the FreeBSD-specific patch, when the environment variable
e2fsprogs_inhibit_SIGINFO exists (whatever its content, even if empty),
do NOT install the SIGINFO handler.  Leverage this when running the
self-tests.  This is to avoid false negatives during the self-tests due
to interspersed SIGINFO output redirected from stderr to the log files.

Bump PORTREVISION because this changes the e2fsck utility.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@421726 35697150-7ecd-e111-bb59-0022644237b5